move types to flows/CostCalculator
